+++ OpenGL/OpenGL/src/Texture.h
@@ -13,7 +13,7 @@ private:
 	int m_Width, m_Height, m_BPP;
 public:
 	Texture(const std::string& path);
-	Texture(const std::string& path, std::map<GLenum, GLfloat> options);
+	Texture(const std::string& path, std::map<GLenum, GLint> options);
 	~Texture();
 
 	void Bind(unsigned int slot = 0) const;
